+++ CHANGELOG.md
@@ -40,7 +40,7 @@ Releasing is documented in RELEASE.md
 ### Removed
 
 ### Fixed
-
+- fix http pooling for matrix generation requests
 ### Security